fix: avatar name address truncation
