+++ packages/ui/src/components/avatar_name/styles.ts
@@ -3,6 +3,7 @@ import { makeStyles } from 'tss-react/mui';
 const useStyles = makeStyles()((theme) => ({
   root: {
     display: 'flex',
+    flex: 1,
     alignItems: 'center',
     justifyContent: 'flex-start',
     gap: theme.spacing(1),
@@ -12,15 +13,11 @@ const useStyles = makeStyles()((theme) => ({
     },
   },
   avatar: {
-    '&': {
-      flex: `0 0 ${theme.spacing(3.5)}`,
-    },
+    flex: `0 0 ${theme.spacing(3.5)}`,
   },
   text: {
-    '&': {
-      flex: `1 1 auto`,
-      color: theme.palette.custom.fonts.highlight,
-    },
+    flex: `1`,
+    color: theme.palette.custom.fonts.highlight,
   },
   popper: {
     marginTop: `-${theme.spacing(2)} !important`,
